refactor: replace ws with rpc
