+++ src/main/java/com/limechain/Main.java
@@ -1,12 +1,12 @@
 package com.limechain;
 
 import com.limechain.lightclient.LightClient;
-import com.limechain.rpc.ws.server.WebSocketRPC;
+import com.limechain.rpc.server.RpcApp;
 import sun.misc.Signal;
 
 public class Main {
     public static void main(String[] args) {
-        WebSocketRPC wsRpc = new WebSocketRPC();
+        RpcApp wsRpc = new RpcApp();
         LightClient client = new LightClient(args, wsRpc);
 
         client.start();
