+++ src/main/java/com/limechain/rpc/client/SubscriptionRpcClient.java
@@ -1,4 +1,4 @@
-package com.limechain.rpc.ws.client;
+package com.limechain.rpc.client;
 
 import com.limechain.rpc.pubsub.Message;
 import com.limechain.rpc.pubsub.PubSubService;
