+++ src/main/java/com/limechain/rpc/server/AppBean.java
@@ -1,4 +1,4 @@
-package com.limechain.rpc.ws.server;
+package com.limechain.rpc.server;
 
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
